Cleanup, remove warnings, more checking in placeholder, merging commons
Check for fork() failing (can happen).  Merge the two execve() points for
the original app_process32 together.  Make the daemon call an execve().

Move the declaration of setup_policy() to a header.  Avoids problems if
it gets changed somewhere.

Adjust the use of types in policy.c, while hashtab_key_t is a char *, the
compiler emits warnings if you're not calling it that.  Also add const to
these since the value IS constant.