Add initial 64-bit mode support
This builds 64-bit versions of the binaries.  The 64-bit versions expect
to chain to app_process64.  This adds another hook into the system making
us less likely to get expelled.  I suspect on 64-bit systems
app_process32 may never get called in a privledged context.

The update script needs to be modified to add in the extra hook.